+++ source/app/transform/transforms/knntransform.cpp
@@ -64,7 +64,8 @@ void KNNTransform::apply(TransformedGraph& target) const
             removees.set(*it, false);
         }
 
-        target.setProgress((progress++ * 100u) / static_cast<uint64_t>(target.numNodes()));
+        target.setProgress(static_cast<int>((progress++ * 100u) /
+            static_cast<uint64_t>(target.numNodes())));
     }
 
     progress = 0;
@@ -87,7 +88,8 @@ void KNNTransform::apply(TransformedGraph& target) const
                 rank._mean = static_cast<double>(rank._source + rank._target) * 0.5;
         }
 
-        target.setProgress((progress++ * 100u) / static_cast<uint64_t>(target.numEdges()));
+        target.setProgress(static_cast<int>((progress++ * 100u) /
+            static_cast<uint64_t>(target.numEdges())));
     }
 
     target.setProgress(-1);
