+++ source/app/transform/transforms/spanningtreetransform.cpp
@@ -76,7 +76,8 @@ void SpanningTreeTransform::apply(TransformedGraph& target) const
         if(removees.get(edgeId))
             target.mutableGraph().removeEdge(edgeId);
 
-        target.setProgress((progress++ * 100u) / static_cast<uint64_t>(target.numEdges()));
+        target.setProgress(static_cast<int>((progress++ * 100u) /
+            static_cast<uint64_t>(target.numEdges())));
     }
 
     target.setProgress(-1);
